Add possibility to specify channels in parameters
Using the default list of channels is not really intuitive
and should be discouraged in exchange for sending the channel
list with each request to the bsread writer. For now the
public API remains the same, but the default channels list
will be removed in the future.

The user can set the list of channels he wants to retrieve by
setting the attribute "channels" (as a list of channels) in
the configuration of the writer.